Refactored App component into multiple files
